Add files via upload
Different age groups experience different types of crimes. Young people often face violent crimes, while older people are more likely to experience property crimes. These patterns can vary based on factors like location and income. Understanding these crime patterns is crucial for creating better strategies to prevent crime, support victims, and address the significant impact of fear on community behavior and lifestyle choices

The data for this project was primarily obtained from the CMPD Data provided to the students. The project also incorporates Census Data to study crime patterns. The CMPD Data, spanning from 2012 to 2016, was cleaned and merged to ensure data integrity. This process was not without its challenges, as it involved learning how to work with dictionaries and dealing with issues such as duplicate columns. Initial mistakes were made in concatenating the data frames which required a restart from scratch. To maintain simplicity in the data, only the first case of any duplicate cases or cases with multiple victims was included. In terms of data merging, specific columns- namely Incident, Victim, Offense, and Weapons, from the CMPD Data were joined with the Census Data. This merging was done based on the Complain_No Column to ensure consistency of information across the dataset and to avoid recurring columns. This rigorous process of data cleaning and merging has laid a solid foundation for the analysis of crime patterns in the following stages of the project.